+++ src/main/java/results/repository/SubjectRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.repository.query.Param;
 import results.model.Student;
 import results.model.Subject;
 
-public interface SubjectRepository extends CrudRepository<Subject, Long> {
+public interface SubjectRepository extends CrudRepository<Subject, Integer> {
     Page<Subject> findAll(Pageable pageable);
     @Query("FROM Subject g where g.classArm.armId= :armId")
    Iterable<Subject> findAllByClassArm(@Param("armId") Long armId);
